requirements.txt
click
rich
beautifulsoup4
requests
websocket-client
seleniumbase
